Moved DifferenceElement classes and static methods to new class DifferenceElementCalculator
