+++ javaparser-core/src/main/java/com/github/javaparser/printer/lexicalpreservation/Difference.java
@@ -4,7 +4,6 @@ import com.github.javaparser.GeneratedJavaParserConstants;
 import com.github.javaparser.TokenTypes;
 import com.github.javaparser.ast.Node;
 import com.github.javaparser.ast.comments.Comment;
-import com.github.javaparser.ast.type.PrimitiveType;
 import com.github.javaparser.printer.concretesyntaxmodel.*;
 import com.github.javaparser.printer.lexicalpreservation.LexicalDifferenceCalculator.CsmChild;
 
@@ -22,474 +21,10 @@ public class Difference {
 
     private static final int STANDARD_INDENTATION_SIZE = 4;
 
-    private final List<DifferenceElement> elements;
+    private final List<DifferenceElementCalculator.DifferenceElement> elements;
 
-    private Difference(List<DifferenceElement> elements) {
+    Difference(List<DifferenceElementCalculator.DifferenceElement> elements) {
         this.elements = elements;
-    }
-
-    interface DifferenceElement {
-        static DifferenceElement added(CsmElement element) {
-            return new Added(element);
-        }
-
-        static DifferenceElement removed(CsmElement element) {
-            return new Removed(element);
-        }
-
-        static DifferenceElement kept(CsmElement element) {
-            return new Kept(element);
-        }
-
-        /**
-         * Return the CsmElement considered in this DifferenceElement.
-         */
-        CsmElement getElement();
-
-        boolean isAdded();
-    }
-
-    private static class Added implements DifferenceElement {
-        final CsmElement element;
-
-        Added(CsmElement element) {
-            this.element = element;
-        }
-
-        @Override
-        public String toString() {
-            return "Added{" + element + '}';
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-
-            Added added = (Added) o;
-
-            return element.equals(added.element);
-        }
-
-        @Override
-        public int hashCode() {
-            return element.hashCode();
-        }
-
-        @Override
-        public CsmElement getElement() {
-            return element;
-        }
-
-        @Override
-        public boolean isAdded() {
-            return true;
-        }
-
-        boolean isIndent() { return element instanceof CsmIndent; }
-
-        boolean isUnindent() { return element instanceof CsmUnindent; }
-
-        TextElement toTextElement() {
-            if (element instanceof CsmChild) {
-                return new ChildTextElement(((CsmChild) element).getChild());
-            } else if (element instanceof CsmToken) {
-                return new TokenTextElement(((CsmToken) element).getTokenType(), ((CsmToken) element).getContent(null));
-            } else {
-                throw new UnsupportedOperationException(element.getClass().getSimpleName());
-            }
-        }
-    }
-
-    /**
-     * Elements in a CsmMix have been reshuffled. It could also mean that
-     * some new elements have been added or removed to the mix.
-     */
-    private static class Reshuffled implements DifferenceElement {
-        final CsmMix previousOrder;
-        final CsmMix element;
-
-        Reshuffled(CsmMix previousOrder, CsmMix element) {
-            this.previousOrder = previousOrder;
-            this.element = element;
-        }
-
-        @Override
-        public String toString() {
-            return "Reshuffled{" + element + ", previous="+ previousOrder+ '}';
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-
-            Reshuffled that = (Reshuffled) o;
-
-            if (!previousOrder.equals(that.previousOrder)) return false;
-            return element.equals(that.element);
-        }
-
-        @Override
-        public int hashCode() {
-            int result = previousOrder.hashCode();
-            result = 31 * result + element.hashCode();
-            return result;
-        }
-
-        @Override
-        public CsmMix getElement() {
-            return element;
-        }
-
-        @Override
-        public boolean isAdded() {
-            return false;
-        }
-    }
-
-    private static class Kept implements DifferenceElement {
-        final CsmElement element;
-
-        Kept(CsmElement element) {
-            this.element = element;
-        }
-
-        @Override
-        public String toString() {
-            return "Kept{" + element + '}';
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-
-            Kept kept = (Kept) o;
-
-            return element.equals(kept.element);
-        }
-
-        @Override
-        public int hashCode() {
-            return element.hashCode();
-        }
-
-        @Override
-        public CsmElement getElement() {
-            return element;
-        }
-
-        int getTokenType() {
-            if (isToken()) {
-                CsmToken csmToken = (CsmToken) element;
-                return csmToken.getTokenType();
-            }
-
-            throw new IllegalStateException("Kept is not a " + CsmToken.class.getSimpleName());
-        }
-
-        @Override
-        public boolean isAdded() {
-            return false;
-        }
-
-        boolean isIndent() { return element instanceof CsmIndent; }
-
-        boolean isUnindent() { return element instanceof CsmUnindent; }
-
-        boolean isToken() { return element instanceof CsmToken; }
-
-        boolean isChild() { return element instanceof CsmChild; }
-
-        boolean isPrimitiveType() {
-            if (isChild()) {
-                CsmChild csmChild = (CsmChild) element;
-                return csmChild.getChild() instanceof PrimitiveType;
-            }
-
-            return false;
-        }
-
-        boolean isWhiteSpace() {
-            if(isToken()) {
-                CsmToken csmToken = (CsmToken) element;
-                return csmToken.isWhiteSpace();
-            }
-
-            return false;
-        }
-
-        boolean isWhiteSpaceOrComment() {
-            if (isToken()) {
-                CsmToken csmToken = (CsmToken) element;
-                return TokenTypes.isWhitespaceOrComment(csmToken.getTokenType());
-            }
-
-            return false;
-        }
-    }
-
-    private static class Removed implements DifferenceElement {
-        final CsmElement element;
-
-        Removed(CsmElement element) {
-            this.element = element;
-        }
-
-        @Override
-        public String toString() {
-            return "Removed{" + element + '}';
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-
-            Removed removed = (Removed) o;
-
-            return element.equals(removed.element);
-        }
-
-        @Override
-        public int hashCode() {
-            return element.hashCode();
-        }
-
-        @Override
-        public CsmElement getElement() {
-            return element;
-        }
-
-        Node getChild() {
-            if (isChild()) {
-                CsmChild csmChild = (CsmChild) element;
-                return csmChild.getChild();
-            }
-
-            throw new IllegalStateException("Removed is not a " + CsmChild.class.getSimpleName());
-        }
-
-        int getTokenType() {
-            if (isToken()) {
-                CsmToken csmToken = (CsmToken) element;
-                return csmToken.getTokenType();
-            }
-
-            throw new IllegalStateException("Removed is not a " + CsmToken.class.getSimpleName());
-        }
-
-        @Override
-        public boolean isAdded() {
-            return false;
-        }
-
-        boolean isToken() { return element instanceof CsmToken; }
-
-        boolean isChild() { return element instanceof CsmChild; }
-
-        boolean isPrimitiveType() {
-            if (isChild()) {
-                CsmChild csmChild = (CsmChild) element;
-                return csmChild.getChild() instanceof PrimitiveType;
-            }
-
-            return false;
-        }
-
-        boolean isWhiteSpace() {
-            if(isToken()) {
-                CsmToken csmToken = (CsmToken) element;
-                return csmToken.isWhiteSpace();
-            }
-
-            return false;
-        }
-    }
-
-    private static boolean matching(CsmElement a, CsmElement b) {
-        if (a instanceof CsmChild) {
-            if (b instanceof CsmChild) {
-                CsmChild childA = (CsmChild) a;
-                CsmChild childB = (CsmChild) b;
-                return childA.getChild().equals(childB.getChild());
-            } else if (b instanceof CsmToken) {
-                return false;
-            } else if (b instanceof CsmIndent) {
-                return false;
-            } else if (b instanceof CsmUnindent) {
-                return false;
-            } else {
-                throw new UnsupportedOperationException(a.getClass().getSimpleName()+ " "+b.getClass().getSimpleName());
-            }
-        } else if (a instanceof CsmToken) {
-            if (b instanceof CsmToken) {
-                CsmToken childA = (CsmToken)a;
-                CsmToken childB = (CsmToken)b;
-                return childA.getTokenType() == childB.getTokenType();
-            } else if (b instanceof CsmChild) {
-                return false;
-            } else if (b instanceof CsmIndent) {
-                return false;
-            } else if (b instanceof CsmUnindent) {
-                return false;
-            } else {
-                throw new UnsupportedOperationException(a.getClass().getSimpleName()+ " "+b.getClass().getSimpleName());
-            }
-        } else if (a instanceof CsmIndent) {
-            return b instanceof CsmIndent;
-        } else if (a instanceof CsmUnindent) {
-            return b instanceof CsmUnindent;
-        }
-        throw new UnsupportedOperationException(a.getClass().getSimpleName()+ " "+b.getClass().getSimpleName());
-    }
-
-    private static boolean replacement(CsmElement a, CsmElement b) {
-        if (a instanceof CsmIndent || b instanceof CsmIndent || a instanceof CsmUnindent || b instanceof CsmUnindent) {
-            return false;
-        }
-        if (a instanceof CsmChild) {
-            if (b instanceof CsmChild) {
-                CsmChild childA = (CsmChild) a;
-                CsmChild childB = (CsmChild) b;
-                return childA.getChild().getClass().equals(childB.getChild().getClass());
-            } else if (b instanceof CsmToken) {
-                return false;
-            } else {
-                throw new UnsupportedOperationException(a.getClass().getSimpleName()+ " "+b.getClass().getSimpleName());
-            }
-        } else if (a instanceof CsmToken) {
-            if (b instanceof CsmToken) {
-                CsmToken childA = (CsmToken)a;
-                CsmToken childB = (CsmToken)b;
-                return childA.getTokenType() == childB.getTokenType();
-            } else if (b instanceof CsmChild) {
-                return false;
-            }
-        }
-        throw new UnsupportedOperationException(a.getClass().getSimpleName()+ " "+b.getClass().getSimpleName());
-    }
-
-    /**
-     * Find the positions of all the given children.
-     */
-    private static Map<Node, Integer> findChildrenPositions(LexicalDifferenceCalculator.CalculatedSyntaxModel calculatedSyntaxModel) {
-        Map<Node, Integer> positions = new HashMap<>();
-        for (int i=0;i<calculatedSyntaxModel.elements.size();i++) {
-            CsmElement element = calculatedSyntaxModel.elements.get(i);
-            if (element instanceof CsmChild) {
-                positions.put(((CsmChild)element).getChild(), i);
-            }
-        }
-        return positions;
-    }
-
-    /**
-     * Calculate the Difference between two CalculatedSyntaxModel elements, determining which elements were kept,
-     * which were added and which were removed.
-     */
-    static Difference calculate(LexicalDifferenceCalculator.CalculatedSyntaxModel original, LexicalDifferenceCalculator.CalculatedSyntaxModel after) {
-        // For performance reasons we use the positions of matching children
-        // to guide the calculation of the difference
-        //
-        // Suppose we have:
-        //   qwerty[A]uiop
-        //   qwer[A]uiop
-        //
-        // with [A] being a child and lowercase letters being tokens
-        //
-        // We would calculate the Difference between "qwerty" and "qwer" then we know the A is kept, and then we
-        // would calculate the difference between "uiop" and "uiop"
-
-        Map<Node, Integer> childrenInOriginal = findChildrenPositions(original);
-        Map<Node, Integer> childrenInAfter = findChildrenPositions(after);
-
-        List<Node> commonChildren = new LinkedList<>(childrenInOriginal.keySet());
-        commonChildren.retainAll(childrenInAfter.keySet());
-        commonChildren.sort(Comparator.comparingInt(childrenInOriginal::get));
-
-        List<DifferenceElement> elements = new LinkedList<>();
-
-        int originalIndex = 0;
-        int afterIndex = 0;
-        int commonChildrenIndex = 0;
-        while (commonChildrenIndex < commonChildren.size()) {
-            Node child = commonChildren.get(commonChildrenIndex++);
-            int posOfNextChildInOriginal = childrenInOriginal.get(child);
-            int posOfNextChildInAfter    = childrenInAfter.get(child);
-            if (originalIndex < posOfNextChildInOriginal || afterIndex < posOfNextChildInAfter) {
-                elements.addAll(calculateImpl(original.sub(originalIndex, posOfNextChildInOriginal), after.sub(afterIndex, posOfNextChildInAfter)).elements);
-            }
-            elements.add(new Kept(new CsmChild(child)));
-            originalIndex = posOfNextChildInOriginal + 1;
-            afterIndex = posOfNextChildInAfter + 1;
-        }
-
-        if (originalIndex < original.elements.size() || afterIndex < after.elements.size()) {
-            elements.addAll(calculateImpl(original.sub(originalIndex, original.elements.size()), after.sub(afterIndex, after.elements.size())).elements);
-        }
-        return new Difference(elements);
-    }
-
-    private static Difference calculateImpl(LexicalDifferenceCalculator.CalculatedSyntaxModel original, LexicalDifferenceCalculator.CalculatedSyntaxModel after) {
-        List<DifferenceElement> elements = new LinkedList<>();
-
-        int originalIndex = 0;
-        int afterIndex = 0;
-
-        // We move through the two CalculatedSyntaxModel, moving both forward when we have a match
-        // and moving just one side forward when we have an element kept or removed
-
-        do {
-            if (originalIndex < original.elements.size() && afterIndex >= after.elements.size()) {
-                elements.add(new Removed(original.elements.get(originalIndex)));
-                originalIndex++;
-            } else if (originalIndex >= original.elements.size() && afterIndex < after.elements.size()) {
-                elements.add(new Added(after.elements.get(afterIndex)));
-                afterIndex++;
-            } else {
-                CsmElement nextOriginal = original.elements.get(originalIndex);
-                CsmElement nextAfter = after.elements.get(afterIndex);
-
-                if ((nextOriginal instanceof CsmMix) && (nextAfter instanceof CsmMix)) {
-                    if (((CsmMix) nextAfter).getElements().equals(((CsmMix) nextOriginal).getElements())) {
-                        // No reason to deal with a reshuffled, we are just going to keep everything as it is
-                        ((CsmMix) nextAfter).getElements().forEach(el -> elements.add(new Kept(el)));
-                    } else {
-                        elements.add(new Reshuffled((CsmMix)nextOriginal, (CsmMix)nextAfter));
-                    }
-                    originalIndex++;
-                    afterIndex++;
-                } else if (matching(nextOriginal, nextAfter)) {
-                    elements.add(new Kept(nextOriginal));
-                    originalIndex++;
-                    afterIndex++;
-                } else if (replacement(nextOriginal, nextAfter)) {
-                    elements.add(new Removed(nextOriginal));
-                    elements.add(new Added(nextAfter));
-                    originalIndex++;
-                    afterIndex++;
-                } else {
-                    // We can try to remove the element or add it and look which one leads to the lower difference
-                    Difference adding = calculate(original.from(originalIndex), after.from(afterIndex + 1));
-                    Difference removing = null;
-                    if (adding.cost() > 0) {
-                        removing = calculate(original.from(originalIndex + 1), after.from(afterIndex));
-                    }
-
-                    if (removing == null || removing.cost() > adding.cost()) {
-                        elements.add(new Added(nextAfter));
-                        afterIndex++;
-                    } else {
-                        elements.add(new Removed(nextOriginal));
-                        originalIndex++;
-                    }
-                }
-            }
-        } while (originalIndex < original.elements.size() || afterIndex < after.elements.size());
-
-        return new Difference(elements);
     }
 
     private List<TextElement> processIndentation(List<TokenTextElement> indentation, List<TextElement> prevElements) {
@@ -577,13 +112,13 @@ public class Difference {
         List<TextElement> originalElements = nodeText.getElements();
         int originalIndex = 0;
 
-        List<DifferenceElement> diffElements = getElements();
+        List<DifferenceElementCalculator.DifferenceElement> diffElements = getElements();
         int diffIndex = 0;
         do {
             if (diffIndex < diffElements.size() && originalIndex >= originalElements.size()) {
-                DifferenceElement diffElement = diffElements.get(diffIndex);
-                if (diffElement instanceof Kept) {
-                    Kept kept = (Kept) diffElement;
+                DifferenceElementCalculator.DifferenceElement diffElement = diffElements.get(diffIndex);
+                if (diffElement instanceof DifferenceElementCalculator.Kept) {
+                    DifferenceElementCalculator.Kept kept = (DifferenceElementCalculator.Kept) diffElement;
 
                     if (kept.isWhiteSpaceOrComment()) {
                         diffIndex++;
@@ -591,8 +126,8 @@ public class Difference {
                         throw new IllegalStateException("Cannot keep element because we reached the end of nodetext: "
                                 + nodeText + ". Difference: " + this);
                     }
-                } else if (diffElement instanceof Added) {
-                    Added addedElement = (Added) diffElement;
+                } else if (diffElement instanceof DifferenceElementCalculator.Added) {
+                    DifferenceElementCalculator.Added addedElement = (DifferenceElementCalculator.Added) diffElement;
 
                     nodeText.addElement(originalIndex, addedElement.toTextElement());
                     originalIndex++;
@@ -610,10 +145,10 @@ public class Difference {
                             + this + " " + originalElement);
                 }
             } else {
-                DifferenceElement diffElement = diffElements.get(diffIndex);
+                DifferenceElementCalculator.DifferenceElement diffElement = diffElements.get(diffIndex);
 
-                if (diffElement instanceof Added) {
-                    Added addedElement = (Added) diffElement;
+                if (diffElement instanceof DifferenceElementCalculator.Added) {
+                    DifferenceElementCalculator.Added addedElement = (DifferenceElementCalculator.Added) diffElement;
 
                     if (addedElement.isIndent()) {
                         for (int i=0;i<STANDARD_INDENTATION_SIZE;i++){
@@ -678,8 +213,8 @@ public class Difference {
                     boolean originalElementIsChild = originalElement instanceof ChildTextElement;
                     boolean originalElementIsToken = originalElement instanceof TokenTextElement;
 
-                    if (diffElement instanceof Kept) {
-                        Kept kept = (Kept)diffElement;
+                    if (diffElement instanceof DifferenceElementCalculator.Kept) {
+                        DifferenceElementCalculator.Kept kept = (DifferenceElementCalculator.Kept)diffElement;
 
                         if (originalElement.isComment()) {
                             originalIndex++;
@@ -723,8 +258,8 @@ public class Difference {
                         } else {
                             throw new UnsupportedOperationException("kept " + kept.element + " vs " + originalElement);
                         }
-                    } else if (diffElement instanceof Removed) {
-                        Removed removed = (Removed)diffElement;
+                    } else if (diffElement instanceof DifferenceElementCalculator.Removed) {
+                        DifferenceElementCalculator.Removed removed = (DifferenceElementCalculator.Removed)diffElement;
 
                         if (removed.isChild() && originalElementIsChild) {
                             ChildTextElement originalElementChild = (ChildTextElement)originalElement;
@@ -742,7 +277,7 @@ public class Difference {
                                 if (originalIndex < originalElements.size() && originalElements.get(originalIndex).isNewline()) {
                                     originalIndex = considerCleaningTheLine(nodeText, originalIndex);
                                 } else {
-                                    if (diffIndex + 1 >= this.getElements().size() || !(this.getElements().get(diffIndex + 1) instanceof Added)) {
+                                    if (diffIndex + 1 >= this.getElements().size() || !(this.getElements().get(diffIndex + 1) instanceof DifferenceElementCalculator.Added)) {
                                         originalIndex = considerEnforcingIndentation(nodeText, originalIndex);
                                     }
                                     // If in front we have one space and before also we had space let's drop one space
@@ -750,7 +285,7 @@ public class Difference {
                                         if (originalElements.get(originalIndex).isWhiteSpace()
                                                 && originalElements.get(originalIndex - 1).isWhiteSpace()) {
                                             // However we do not want to do that when we are about to adding or removing elements
-                                            if ((diffIndex + 1) == diffElements.size() || (diffElements.get(diffIndex + 1) instanceof Kept)) {
+                                            if ((diffIndex + 1) == diffElements.size() || (diffElements.get(diffIndex + 1) instanceof DifferenceElementCalculator.Kept)) {
                                                 originalElements.remove(originalIndex--);
                                             }
                                         }
@@ -778,9 +313,9 @@ public class Difference {
                         } else {
                             throw new UnsupportedOperationException("removed " + removed.element + " vs " + originalElement);
                         }
-                    } else if (diffElement instanceof Reshuffled) {
+                    } else if (diffElement instanceof DifferenceElementCalculator.Reshuffled) {
                         // First, let's see how many tokens we need to attribute to the previous version of the of the CsmMix
-                        Reshuffled reshuffled = (Reshuffled)diffElement;
+                        DifferenceElementCalculator.Reshuffled reshuffled = (DifferenceElementCalculator.Reshuffled)diffElement;
                         CsmMix elementsFromPreviousOrder = reshuffled.previousOrder;
                         CsmMix elementsFromNextOrder = reshuffled.element;
 
@@ -844,16 +379,16 @@ public class Difference {
                                     int indexOfOriginalCSMElement = nodeTextIndexToPreviousCSMIndex.get(ntIndex);
                                     if (elementsToAddBeforeGivenOriginalCSMElement.containsKey(indexOfOriginalCSMElement)) {
                                         for (CsmElement elementToAdd : elementsToAddBeforeGivenOriginalCSMElement.get(indexOfOriginalCSMElement)) {
-                                            diffElements.add(diffElIterator++, new Added(elementToAdd));
+                                            diffElements.add(diffElIterator++, new DifferenceElementCalculator.Added(elementToAdd));
                                         }
                                     }
 
                                     CsmElement originalCSMElement = elementsFromPreviousOrder.getElements().get(indexOfOriginalCSMElement);
                                     boolean toBeKept = correspondanceBetweenNextOrderAndPreviousOrder.containsValue(indexOfOriginalCSMElement);
                                     if (toBeKept) {
-                                        diffElements.add(diffElIterator++, new Kept(originalCSMElement));
+                                        diffElements.add(diffElIterator++, new DifferenceElementCalculator.Kept(originalCSMElement));
                                     } else {
-                                        diffElements.add(diffElIterator++, new Removed(originalCSMElement));
+                                        diffElements.add(diffElIterator++, new DifferenceElementCalculator.Removed(originalCSMElement));
                                     }
                                 }
                                 // else we have a simple node text element, without associated csm element, just keep ignore it
@@ -863,7 +398,7 @@ public class Difference {
                         // Finally we look for the remaining new elements that were not yet added and
                         // add all of them
                         for (CsmElement elementToAdd : elementsToBeAddedAtTheEnd) {
-                            diffElements.add(diffElIterator++, new Added(elementToAdd));
+                            diffElements.add(diffElIterator++, new DifferenceElementCalculator.Added(elementToAdd));
                         }
                     } else {
                         throw new UnsupportedOperationException("" + diffElement + " vs " + originalElement);
@@ -885,7 +420,7 @@ public class Difference {
             for (int pi = 0; pi< previousOrderElements.size() && !found; pi++) {
                 CsmElement pe = previousOrderElements.get(pi);
                 if (!correspondanceBetweenNextOrderAndPreviousOrder.values().contains(pi)
-                        && matching(ne, pe)) {
+                        && DifferenceElementCalculator.matching(ne, pe)) {
                     found = true;
                     correspondanceBetweenNextOrderAndPreviousOrder.put(ni, pi);
                 }
@@ -895,7 +430,7 @@ public class Difference {
         return correspondanceBetweenNextOrderAndPreviousOrder;
     }
 
-    private boolean isFollowedByUnindent(List<DifferenceElement> diffElements, int diffIndex) {
+    private boolean isFollowedByUnindent(List<DifferenceElementCalculator.DifferenceElement> diffElements, int diffIndex) {
         return (diffIndex + 1) < diffElements.size()
                 && diffElements.get(diffIndex + 1).isAdded()
                 && diffElements.get(diffIndex + 1).getElement() instanceof CsmUnindent;
@@ -1009,7 +544,7 @@ public class Difference {
     }
 
     private boolean isAReplacement(int diffIndex) {
-        return (diffIndex > 0) && getElements().get(diffIndex) instanceof Added && getElements().get(diffIndex - 1) instanceof Removed;
+        return (diffIndex > 0) && getElements().get(diffIndex) instanceof DifferenceElementCalculator.Added && getElements().get(diffIndex - 1) instanceof DifferenceElementCalculator.Removed;
     }
 
     private boolean isPrimitiveType(TextElement textElement) {
@@ -1028,8 +563,8 @@ public class Difference {
         }
     }
 
-    private long cost() {
-        return elements.stream().filter(e -> !(e instanceof Kept)).count();
+    long cost() {
+        return elements.stream().filter(e -> !(e instanceof DifferenceElementCalculator.Kept)).count();
     }
 
     @Override
@@ -1037,7 +572,7 @@ public class Difference {
         return "Difference{" + elements + '}';
     }
 
-    List<DifferenceElement> getElements() {
+    List<DifferenceElementCalculator.DifferenceElement> getElements() {
         return elements;
     }
